add business logic and api integratation with firebase in the home page

+++ lib/models/product.dart
@@ -7,7 +7,7 @@ class Product {
   final String imageUrl;
   final int discountValue;
   final String category;
-  final double? rate;
+  final int? rate;
 
   Product({
     required this.id,
@@ -18,6 +18,7 @@ class Product {
     this.category = "Other",
     this.rate,
   });
+
 
   Map<String, dynamic> toMap() {
     return {
@@ -39,7 +40,7 @@ class Product {
       imageUrl: map["imageUrl"] as String,
       discountValue: map["discountValue"] as int,
       category: map["category"] as String,
-      rate: map["rate"] as double,
+      rate: map["rate"] as int,
     );
   }
 }
